Load data as float 	modified:   np.py

+++ np.py
@@ -13,5 +13,5 @@ def print(indata):
     np.savetxt(sys.stdout, indata, fmt='%.18g')
 
 
-data = np.loadtxt(sys.stdin, dtype=np.int)  # See: https://stackoverflow.com/a/8192426/1679629
+data = np.loadtxt(sys.stdin)  # See: https://stackoverflow.com/a/8192426/1679629
 exec(sys.argv[1])
